refactor: renamed modules, optimized regex and several func modifiers
