+++ ethr-status/src/test/java/me/uport/sdk/ethr_status/EthrStatusRegistryTests.kt
@@ -1,13 +1,14 @@
-package me.uport.sdk.ethrstatusregistry
+package me.uport.sdk.ethr_status
 
 import assertk.assertThat
 import assertk.assertions.isFalse
+import kotlinx.coroutines.runBlocking
 import org.junit.Test
 
 class EthrStatusRegistryTests {
 
     @Test
-    fun `can check for revocation`() {
+    fun `can check for revocation`() = runBlocking {
 
         val validShareReqToken =
             "eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NkstUiJ9.eyJjbGFpbXMiOnsibmFtZSI6IlIgRGFuZWVsIE9saXZhdyJ9LCJzdGF0dXMiOnsidHlwZSI6IkV0aHJTdGF0dXNSZWdpc3RyeTIwMTkiLCJpZCI6ImRpZDpldGhyOjB4ZThjOTFiZGU3NjI1YWIyYzBlZDlmMjE0ZGViMzk0NDBkYTdlMDNjNCJ9LCJpYXQiOjEyMzQ1Njc4LCJleHAiOjEyMzQ1OTc4LCJpc3MiOiJkaWQ6ZXRocjoweDQxMjNjYmQxNDNiNTVjMDZlNDUxZmYyNTNhZjA5Mjg2YjY4N2E5NTAifQ.iRuhl5JHPIBLP_WSr3vRwM8MWTKwJUAUk5AzDDuRvyySsQXuZ-W3Ri3H9WISytZjwHy3ZAnBTcZ6_gkTfdvYhAE"
@@ -15,4 +16,13 @@ class EthrStatusRegistryTests {
         val result = statusRegistry.checkStatus(validShareReqToken)
         assertThat(result).isFalse()
     }
+
+    /*@Test
+    fun `can parse valid id`(){
+
+        val statusRegistry = EthrStatusRegistry()
+        val (address, network) = statusRegistry.parseRegistryId("mainnet:0x63142bebe28e663ec4d007cd12ce5e77c37deafc")
+        assertThat(network).isEqualTo("mainnet")
+        assertThat(address).isEqualTo("0x63142bebe28e663ec4d007cd12ce5e77c37deafc")
+    }*/
 }
