+++ settings.gradle
@@ -1,8 +1,9 @@
-include ':web-did', ':credential-status', ':ethr-status-registry'
+include ':web-did'
 include ':uport-did'
 include ':ethr-did'
 include ':universal-did'
 include ':jwt'
 include ':uport-defaults'
-
+include ':credential-status'
+include ':ethr-status'
 include ':jwt-test'
